+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.4 - 2022-01-25
 
+### Added
+
+- Safe multithreading for `database` module
+
 ### Changed
 
 - `logging` module